refactor(db): decompose database.py into coherent modules with DataContext facade
Extract the 1194-line database.py god module into 7 focused modules under src/db/:
- models.py: dataclasses with MonthlyMixin for shared computation
- security.py: PBKDF2 password hashing and SHA-256 email hashing
- connection.py: DB_PATH, get_connection, ensure_db_directory
- schema.py: table creation, seeding, init_db
- store.py: all ~40 CRUD functions and yearly overview aggregation
- demo.py: hardcoded demo data and demo-specific query functions
- facade.py: DataContext class for demo-transparent data access

database.py becomes a backward-compat shim re-exporting everything from src.db.
All existing imports continue working unchanged.

Migrate 7 route GET handlers to use DataContext, eliminating if-demo branching.
Add 3 new test files (test_models, test_security_standalone, test_data_context).

Closes #156
