+++ src/db/__init__.py
@@ -1,0 +1,162 @@
+"""Family Budget database package.
+
+Re-exports all public names for convenient access via `from src.db import X`.
+"""
+
+from .connection import DB_PATH, ensure_db_directory, get_connection
+from .demo import (
+    DEMO_EXPENSES,
+    DEMO_EXPENSES_ADVANCED,
+    DEMO_INCOME,
+    DEMO_INCOME_ADVANCED,
+    get_demo_account_totals,
+    get_demo_accounts,
+    get_demo_category_totals,
+    get_demo_expenses,
+    get_demo_expenses_by_category,
+    get_demo_income,
+    get_demo_total_expenses,
+    get_demo_total_income,
+    get_yearly_overview_demo,
+)
+from .facade import DataContext
+from .models import (
+    Account,
+    Category,
+    Expense,
+    Income,
+    PasswordResetToken,
+    User,
+)
+from .schema import DEFAULT_CATEGORIES, init_db
+from .security import (
+    PBKDF2_ITERATIONS,
+    hash_email,
+    hash_password,
+    verify_password,
+)
+from .store import (
+    _calculate_yearly_overview,
+    add_account,
+    add_category,
+    add_expense,
+    add_income,
+    authenticate_user,
+    create_password_reset_token,
+    create_user,
+    delete_account,
+    delete_all_income,
+    delete_category,
+    delete_expense,
+    ensure_default_categories,
+    get_account_by_id,
+    get_account_totals,
+    get_account_usage_count,
+    get_all_accounts,
+    get_all_categories,
+    get_all_expenses,
+    get_all_income,
+    get_category_by_id,
+    get_category_totals,
+    get_category_usage_count,
+    get_expense_by_id,
+    get_expenses_by_category,
+    get_total_income,
+    get_total_monthly_expenses,
+    get_user_by_email,
+    get_user_by_id,
+    get_user_by_username,
+    get_user_count,
+    get_valid_reset_token,
+    get_yearly_overview,
+    mark_reset_token_used,
+    migrate_user_categories,
+    update_account,
+    update_category,
+    update_expense,
+    update_income,
+    update_last_login,
+    update_user_email,
+    update_user_password,
+)
+
+__all__ = [
+    # connection
+    "DB_PATH",
+    # schema
+    "DEFAULT_CATEGORIES",
+    "DEMO_EXPENSES",
+    "DEMO_EXPENSES_ADVANCED",
+    # demo
+    "DEMO_INCOME",
+    "DEMO_INCOME_ADVANCED",
+    # security
+    "PBKDF2_ITERATIONS",
+    "Account",
+    "Category",
+    # facade
+    "DataContext",
+    "Expense",
+    # models
+    "Income",
+    "PasswordResetToken",
+    "User",
+    "_calculate_yearly_overview",
+    "add_account",
+    "add_category",
+    "add_expense",
+    "add_income",
+    "authenticate_user",
+    "create_password_reset_token",
+    "create_user",
+    "delete_account",
+    "delete_all_income",
+    "delete_category",
+    "delete_expense",
+    "ensure_db_directory",
+    "ensure_default_categories",
+    "get_account_by_id",
+    "get_account_totals",
+    "get_account_usage_count",
+    "get_all_accounts",
+    "get_all_categories",
+    "get_all_expenses",
+    # store (all CRUD + aggregation)
+    "get_all_income",
+    "get_category_by_id",
+    "get_category_totals",
+    "get_category_usage_count",
+    "get_connection",
+    "get_demo_account_totals",
+    "get_demo_accounts",
+    "get_demo_category_totals",
+    "get_demo_expenses",
+    "get_demo_expenses_by_category",
+    "get_demo_income",
+    "get_demo_total_expenses",
+    "get_demo_total_income",
+    "get_expense_by_id",
+    "get_expenses_by_category",
+    "get_total_income",
+    "get_total_monthly_expenses",
+    "get_user_by_email",
+    "get_user_by_id",
+    "get_user_by_username",
+    "get_user_count",
+    "get_valid_reset_token",
+    "get_yearly_overview",
+    "get_yearly_overview_demo",
+    "hash_email",
+    "hash_password",
+    "init_db",
+    "mark_reset_token_used",
+    "migrate_user_categories",
+    "update_account",
+    "update_category",
+    "update_expense",
+    "update_income",
+    "update_last_login",
+    "update_user_email",
+    "update_user_password",
+    "verify_password",
+]
