+++ src/db/schema.py
@@ -1,0 +1,111 @@
+"""Database schema creation and initialization."""
+
+from .connection import ensure_db_directory, get_connection
+
+DEFAULT_CATEGORIES = [
+    ("Bolig", "house"),
+    ("Forbrug", "zap"),
+    ("Transport", "car"),
+    ("Børn", "baby"),
+    ("Mad", "utensils"),
+    ("Forsikring", "shield"),
+    ("Abonnementer", "tv"),
+    ("Opsparing", "piggy-bank"),
+    ("Andet", "more-horizontal"),
+]
+
+
+def _create_tables(cur) -> None:
+    """Create all database tables and indexes if they don't exist."""
+    cur.execute("""
+        CREATE TABLE IF NOT EXISTS users (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            username TEXT NOT NULL UNIQUE COLLATE NOCASE,
+            password_hash TEXT NOT NULL,
+            salt TEXT NOT NULL,
+            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+            last_login TIMESTAMP,
+            email_hash TEXT
+        )
+    """)
+
+    cur.execute("""
+        CREATE TABLE IF NOT EXISTS income (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            user_id INTEGER NOT NULL,
+            person TEXT NOT NULL,
+            amount REAL NOT NULL DEFAULT 0,
+            frequency TEXT NOT NULL DEFAULT 'monthly' CHECK(frequency IN ('monthly', 'quarterly', 'semi-annual', 'yearly')),
+            FOREIGN KEY (user_id) REFERENCES users(id),
+            UNIQUE(user_id, person)
+        )
+    """)
+
+    cur.execute("""
+        CREATE TABLE IF NOT EXISTS categories (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            user_id INTEGER NOT NULL,
+            name TEXT NOT NULL,
+            icon TEXT,
+            UNIQUE(user_id, name)
+        )
+    """)
+    cur.execute("CREATE INDEX IF NOT EXISTS idx_categories_user ON categories(user_id, name)")
+
+    cur.execute("""
+        CREATE TABLE IF NOT EXISTS expenses (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            user_id INTEGER NOT NULL,
+            name TEXT NOT NULL,
+            category TEXT NOT NULL,
+            amount REAL NOT NULL,
+            frequency TEXT NOT NULL CHECK(frequency IN ('monthly', 'quarterly', 'semi-annual', 'yearly')),
+            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+            category_id INTEGER REFERENCES categories(id),
+            account TEXT,
+            months TEXT,
+            FOREIGN KEY (user_id) REFERENCES users(id)
+        )
+    """)
+    cur.execute("CREATE INDEX IF NOT EXISTS idx_expenses_category ON expenses(category_id)")
+
+    cur.execute("""
+        CREATE TABLE IF NOT EXISTS accounts (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            user_id INTEGER NOT NULL,
+            name TEXT NOT NULL,
+            UNIQUE(user_id, name)
+        )
+    """)
+    cur.execute("CREATE INDEX IF NOT EXISTS idx_accounts_user ON accounts(user_id, name)")
+
+    cur.execute("""
+        CREATE TABLE IF NOT EXISTS password_reset_tokens (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            user_id INTEGER NOT NULL,
+            token_hash TEXT NOT NULL,
+            expires_at TIMESTAMP NOT NULL,
+            used INTEGER DEFAULT 0,
+            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+            FOREIGN KEY (user_id) REFERENCES users(id)
+        )
+    """)
+
+
+def _seed_default_categories(cur) -> None:
+    """Insert default categories for the demo user (user_id = 0)."""
+    for name, icon in DEFAULT_CATEGORIES:
+        cur.execute(
+            "INSERT OR IGNORE INTO categories (user_id, name, icon) VALUES (?, ?, ?)",
+            (0, name, icon)
+        )
+
+
+def init_db():
+    """Initialize database with schema and default data."""
+    ensure_db_directory()
+    with get_connection() as conn:
+        cur = conn.cursor()
+        _create_tables(cur)
+        _seed_default_categories(cur)
+        conn.commit()
